cleanup interstitial, remove unused hook
