+++ apps/app/components/welcome/featured/slide.tsx
@@ -1,0 +1,40 @@
+import { motion } from "framer-motion";
+
+interface SlideProps {
+  keyName: string;
+  children: React.ReactNode;
+  flipDirection?: boolean;
+}
+
+const slideVariants = {
+  enter: (direction: number) => ({
+    x: direction > 0 ? 1000 : -1000,
+    opacity: 0,
+  }),
+  center: {
+    zIndex: 1,
+    x: 0,
+    opacity: 1,
+  },
+  exit: (direction: number) => ({
+    zIndex: 0,
+    x: direction < 0 ? 1000 : -1000,
+    opacity: 0,
+  }),
+};
+
+const Slide: React.FC<SlideProps> = ({ keyName, children }) => (
+  <motion.div
+    key={keyName}
+    initial="enter"
+    animate="center"
+    exit="exit"
+    variants={slideVariants}
+    transition={{ duration: 0.3 }}
+    className="max-w-2xl w-full mx-auto p-6 space-y-8"
+  >
+    {children}
+  </motion.div>
+);
+
+export default Slide;
