+++ apps/app/lib/utils.ts
@@ -1,0 +1,2 @@
+export const sleep = (ms: number) =>
+  new Promise(resolve => setTimeout(resolve, ms));
